Basic websockets authentication handling
